backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
biopython==1.81
requests==2.31.0
pandas==2.1.3
python-multipart==0.0.6
metapub==0.5.5
aiofiles==23.2.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
httpx==0.25.2
aiohttp==3.9.0
PyPDF2==3.0.1
pdfplumber==0.10.3
tqdm==4.66.1
retrying==1.3.4
